Address review file preview maintainability
